+++ R/check-cols.R
@@ -9,9 +9,9 @@
 #' the data.
 #' @seealso \code{\link{datacheckr}}
 #' @export
-check_cols <- function(data, colnames = NULL, exclusive = TRUE, ordered = TRUE,
+check_cols <- function(data, colnames = NULL, exclusive = FALSE, ordered = FALSE,
                        data_name = substitute(data)) {
-  data_name <- as.character()
+  data_name <- as.character(data_name)
   data <- check_data_frame(data, data_name = data_name)
   data <- check_colnames(data, colnames = colnames, exclusive = exclusive,
                          ordered = ordered, data_name = data_name)