+++ R/check-data-values.R
@@ -12,8 +12,8 @@ check_data_values_column <- function(column_name, data, values, data_name) {
 }
 
 check_data_values <- function(data, values, data_name) {
-  if(is.null(values))
-    invisible(data)
+  if (is.null(values))
+    return(invisible(data))
   column_names <- unique(names(values))
   vapply(column_names, FUN = check_data_values_column, logical(1), data = data, values = values,
          data_name = data_name)