+++ NEWS.md
@@ -3,6 +3,7 @@
 
 - Added function `check_data_frame` to check if an object is a data frame.
 - Added function `check_rows` to check the number of rows in a data frame.
+- Added function `check_cols` to check the names of columns in a data frame.
 - Added function `check_key` to check that particular columns represent unique keys.
 
 - `check_data` now has argument `unique = FALSE` to enforce that columns can only have